Create SearchExtension.cs
WIP #2

Helper method that allows to search for a location based on it's name. Helpful for debugging, as it's easier to find locations based on their name than by sifting through the API. Also possibly helpful if any future search functionality is to be implemented. Could be marked as debug only in this situation but chose not to as it could technically be useful for assignment 2