Rewrote the code for maintainability
Fixes #9

Final fixes, made the code more maintainable at the cost of some performance by creating a Data and Dataset class that hold information. Some information might not be required for every situation, one way to fix this might be through having multiple interfaces that lead down inheritence in ways that group issues together.

This was not done as it seemed unnecessary for the task. If a large amount of datasets are loaded at once and they are expanded upon more, this could lead to some issues, but I interpreted it as downloads being done on a per-need basis instead of in bulk.